+++ src/main/java/com/seregamazur/pulse/cart/dto/CartItemRequest.java
@@ -2,7 +2,7 @@ package com.seregamazur.pulse.cart.dto;
 
 import java.util.UUID;
 
-public record CartItemRequest(UUID productId, int quantity) {
+public record CartItemRequest(UUID productId, long quantity) {
 
 }
